Backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
fastapi-cors
